update utilities, move digitize_labels to utilities
