+++ utilities.py
@@ -1,7 +1,7 @@
-import pandas as pd
 import uuid
 import os
 from collections import defaultdict
+import numpy as np
 
 
 def recursivedict():
@@ -18,23 +18,33 @@ def check_or_create_dir(path):
 
 
 def get_outdir_path(settings, gmt=None):
-    dset_name = self.s["dataset"]["name"]
-    scheme_name = self.s["processing_scheme"]["name"]
-    subset_name = self.s["processing_scheme"]["subset"]
-    est_name = self.s["estimator"]["name"]
+    dset_name = settings["dataset"]["name"]
+    scheme_name = settings["processing_scheme"]["name"]
+    subset_name = settings["processing_scheme"]["subset"]
+    est_name = settings["estimator"]["name"]
     prediction_type = "background_predictions"
     if gmt:
         prediction_type = "geneset_predictions"
     outdir_path = (dset_name + '/' +
-                    scheme_name + '/' + subset_name + '/' +
-                    est_name + '/' +
-                    prediction_type + '/')
+                   scheme_name + '/' + subset_name + '/' +
+                   est_name + '/' +
+                   prediction_type + '/')
     return outdir_path
 
-def get_outfile_name(settings, gmt=None):
+
+def get_outfile_name(gmt=None):
     if gmt:
         outfile_name = gmt.suffix
     else:
         outfile_name = str(uuid.uuid4())
     outfile_name = outfile_name + '.csv'
     return outfile_name
+
+
+def digitize_labels(y, transform_thresholds):
+    """ Take Braak scores and digitize them such that:
+    [0, 1, 2, 3] -> 0
+    [4, 5, 6]    -> 1
+    """
+    y_bin = np.digitize(y, transform_thresholds) - 1
+    return y_bin
